requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
WTForms==3.0.1
nltk==3.7
beautifulsoup4==4.11.1
requests==2.28.2
pyttsx3==2.90
python-pptx==0.6.21
reportlab==3.6.0
pytesseract==0.3.10
Pillow==9.2.0
gunicorn==20.1.0  # Only include this if deploying in production
